backend/requirement.txt
fastapi
uvicorn
faiss-cpu
torch
transformers
Pillow
